+++ src/routes/(admin)/admin/weeklysheet/[sheetId]/+page.svelte
@@ -1,0 +1,29 @@
+<script lang="ts">
+	import type { WeeklySheet } from '$lib/stores/ObjectStore';
+	import type { DocumentReference } from 'firebase/firestore';
+
+	import WeeklySheetForm from '$lib/components/WeeklySheetForm.svelte';
+
+	interface Props {
+		data: {
+			sheetItem: WeeklySheet;
+			docRef: DocumentReference;
+		};
+	}
+
+	let { data }: Props = $props();
+	console.log(data);
+	const thisSheet = data.sheetItem;
+
+	const handleSaveItem = (selectedPDF, newItem) => {
+		console.log('Save called');
+	};
+
+	const handleCancel = () => {
+		console.log('Cancelled called');
+	};
+</script>
+
+<div>
+	<WeeklySheetForm onSubmit={handleSaveItem} onCancel={handleCancel} item={thisSheet} />
+</div>