+++ changelog.md
@@ -16,11 +16,12 @@ Patch: Everything else → bumps patch version (1.0.0 → 1.0.1)
 
 ### Changed
 
--
+- Weekly Sheets now have a publish and unpublish date so they can be available beyond their Sunday
+- The weekly sheets are now loaded correctly by publish and unpublishdate (and Sunday date)
 
 ### Fixed
 
--
+- Fixed the overall management of PDFs throughout the app
 
 ## [5.1.0] - 2025-09-28